Fix bugs in re-centering
There was a bug in re-centering that caused the wrong components of models
to be combined. I'm surprised that this even worked before!
This commit also adds a check to make sure that a source hasn't been
shifted out of the image and re-centers the image in it's initial
position if it has.