Implement S normalization
While the default normalization is still to normalize
the A matrix to unity, it is now possible to normalize
the S matrix to unity instead, or to normalize the S matrix
such that the maximum value is always 1, which *might* be
a more stable normalization for the morphology.
Further testing is needed.