Allow for model updates after A and S updates
Currently the model is only generated once per iteration,
and the same model is used for both the A and S updates.
This commit adds a configuration parameter that allows
the user to update the model twice per iteration, once
after the A update and once after the S update.