Add pytorch and FFT convolutions and translations
Anticipating the upgrade to pytorch to make use of autograd,
we implement FFT to perform PSF convolutions and translations
by fractional amounts. This is not implemented in scarlet yet
but exists in the subdirectory `torch`.

The `torch/utils.py` module contains numpy functionality that is
currently not ported to pytorch yet. The `torch/filters.py` module
contains code to perform FFT convolutions using various sinc
approximations. If we end up using a numpy autograd function or
tensorflow then a new file `filters.py` should be created in either
a `numpy` or `tensorflow` directory with the same API.