Update with Peter's API changes
This updates the new version of scarlet with the API
and architectural changes suggested by @pmelchior.
There are still a few outstanding issues before this branch is
ready for review, namely testing PointSource and MultiComponentSource,
the addition of comments and docstrings, and tests
(although new tests will likely be in a separate branch to allow
Remy and Francois to begin working off of this branch).