Fix proximal operators
Prox weighted was not properly updating after `Component.morph`
was changed to a numpy view into the Tensor, so a copy is made
instead. `uncentered_operator` was fixed to work with even input
morphologies, and `prox_uncentered_symmetry` was updated to allow
the user to fill in the unsymmetric region with zeros.