implement numpy prod workaround
autograd has not yet implemented `np.prod`, so attempts to
use `np.prod(*images)` in `fft_convolve` when one or more of the
images are `ArrayBox`es, the autograd wrapper for arrays used in
gradient calculations, results in an error. This ticket checks
for an `ArrayBox` in `images` and uses a python loop to calculate
the product if one is present.