Convert test data into single numpy archives
The three datasets were converted from directories with
multiple files into three numpy archives that contain all of the
image, PSF, and catalog data for each dataset. This makes them
all easier to load, eliminates the need to use astropy for loading
the test data, and allows them to be used interchangeably in
code.