Do not close figures in scarlet.display methods
When the figures are closed in functions like `show_scene` it prevents
displaying multiple scenes in a loop, making the functionality much
less useful. The problem with not closing the figure is that `fig` is
returned from the functions, causing Jyputer to display them twice.
This can be fixed by adding a `plt.show` in the Jupyter notebook,
giving the user the option to close the scene if they wish or
annotate it if they so desire, making the functions much more
useful with a small change.