Bug fix in PSFDiffKernel
During testing I was surprised to find simultaneously fitting the
deconvolution kernel in each band, even though there are no cross-band
models, is still faster than fitting the kernel in each band separately.
In making the change there was one place in the model that I neglected to
copy properly into the repo, so this commit fixes that bug in the code
and the tutorial.