Make channels necessary and remove in_frame from render
Per the code review we decided to make channels necessary to avoid
unexpcted behavior when using multiple observations that do not
have the same spectral dimension.

This commit also removes `in_frame` from `Observation.render` and
instead uses `Observation.convolve` with a new `convolution_type`
parameter to convolve a model with the PSF that is not the same
shape as the model_frame.