Update Travis script to fix broken tests
This commit fixes tests that were failing in Travis.
By the looks of it conda pulls astropy from an `anaconda`
repository and they changed the version, which used to include
`pytest` as a dependency. They no longer do, causing our test
environment to not install pytest through conda and instead
fell back to the system version of pytest.

This also updates the travis script based on the current
recommendation from conda at
https://docs.conda.io/projects/conda/en/latest/user-guide/tasks/use-conda-with-travis-ci.html