Multiple measurements per page working
This is the last commit with all of the measurements working.
THe problem with this implementation is that for the larger datasets,
eg dataset 1 has 100 sources, the 12 scatter plots contain a total
of 1200 points, which causes the initial load time to be quite large.

So I wanted to save my progress here in case I want to return to this
method of organization, but for ease of use the next several commits
plan to go back to a single measurement per screen.