Fix bug in branch names
This commit fixes an issue that was created when porting the tests
from scarlet_extensions to scarlet. Namely, I didn't want to have
to introduce the `git` package as a dependency so I attempted to
replicate the code on my own. That did not go so well...
So this commit reimplements the git solution that was working.