Various small improvements
@pmelchior noticed that the major difference in runtime between
ADAM and FISTA is the iteration over the proximal operators in
each iteration of the fit. So changing this to 1 or 2
improves the runtime by factors of several while having little
effect on the overall results.

It also creates the detection image in `init_all_sources_main`
to prevent the user from having to do so.